[Proposal] Add requirement for message when publishing a proposal version
The change requires a user to include a message when publishing a new version of a proposal. This message is expected to detail the changes that were made in the new version. The reason for this update is to ensure every version of a proposal has descriptive information on what was changed and why. This will make tracking changes easier in the long run.