Corrected propagation level in Transactional annotation
The propagation level in the Transactional annotation of the  setProcessingErrorMessage method has been updated from REQUIRES_NEW to the default one. This change was done because there's no need to always use a new transaction for setting processing errors, instead it should be part of the current transaction to maintain consistency.