Refine BirdEye customer validation logic in BirdEyeService
Extended the validation check for the BirdEyeCustomer object in the BirdEyeService. Now not only does it check for null customer but also checks if customer's CustomFields is null. The change was necessary since a valid BirdEyeCustomer object could potentially have null CustomFields which would later cause NullPointerException in the application flow. This added validation would mitigate undesired crashes and improve code robustness.