Remove Params and ImmutableMap class usage
Removed Params.java file and ImmutableMap reference in favor of java.util.Map. This change was made to remove the unnecessary dependencies and use standard java classes to provide a more universally understood codebase. It simplifies the codebase by removing the need to understand what Params.java and ImmutableMap is doing, and makes it easier for new developers to learn the codebase as they are more likely to be familiar with standard java classes.