Update library versions in pom.xml
This commit updates the versions of several libraries in the pom.xml file. The changes include upgrades to the Spring Boot parent version, the Springdoc OpenAPI starter, the org.json library, the GraalVM JavaScript engine, and the versions Maven plugin. This ensures that the project is using the latest stable releases of these dependencies.